+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Added
 -----
 - Read version from bot
-- stats: add lexical query to history
+- stats: add lexical query to several stats
 
 Removed
 -------